requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.104.0